added auth and resp check

+++ app/src/main/java/hexlet/code/controller/UserController.java
@@ -3,10 +3,12 @@ package hexlet.code.controller;
 import hexlet.code.dto.UserDTO;
 import hexlet.code.dto.UserCreateDTO;
 import hexlet.code.dto.UserUpdateDTO;
-import hexlet.code.mapper.UserMapper;
 import hexlet.code.service.UserService;
+import hexlet.code.utils.UserUtils;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -28,7 +30,7 @@ public class UserController {
     private UserService userService;
 
     @Autowired
-    private UserMapper userMapper;
+    private UserUtils userUtils;
 
     @GetMapping
     public List<UserDTO> getAllUsers() {
@@ -47,13 +49,19 @@ public class UserController {
     }
 
     @PutMapping("/{id}")
+    @PreAuthorize("authentication.name == userService.findById(#id).email")
     public UserDTO updateUser(@PathVariable Long id, @RequestBody @Valid UserUpdateDTO userUpdateDTO) {
         return userService.update(userUpdateDTO, id);
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteUser(@PathVariable Long id) {
+    //@PreAuthorize("authentication.name == userService.findById(#id).email") что то не то, не сработало
+    public ResponseEntity deleteUser(@PathVariable Long id) {
+        if (!userUtils.getCurrentUser().getEmail().equals(userService.findById(id).getEmail())) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         userService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 }
